refactor: move router in separate file from main, add index for pages
